skyrocket-ai-pipeline/requirements.txt
# Core app framework
streamlit
python-dotenv

# Data processing
pandas==2.2.2
scikit-learn

# Visualization
matplotlib
seaborn

# NLP + AI
nltk
pyspellchecker
sentence-transformers
transformers
torch
openai
